Split partition.hpp header and rename implementation files
Split the partition.hpp header into separate headers for
active and passive partitions, and rename the corresponding
.cpp files for greater uniformity.
