+++ libvast/src/system/index.cpp
@@ -37,10 +37,11 @@
 #include "vast/io/save.hpp"
 #include "vast/logger.hpp"
 #include "vast/partition_synopsis.hpp"
+#include "vast/system/active_partition.hpp"
 #include "vast/system/evaluator.hpp"
 #include "vast/system/local_segment_store.hpp"
 #include "vast/system/meta_index.hpp"
-#include "vast/system/partition.hpp"
+#include "vast/system/passive_partition.hpp"
 #include "vast/system/query_supervisor.hpp"
 #include "vast/system/shutdown.hpp"
 #include "vast/system/status.hpp"
