+++ libvast/test/flatbuffers.cpp
@@ -19,9 +19,10 @@
 #include "vast/msgpack_table_slice.hpp"
 #include "vast/msgpack_table_slice_builder.hpp"
 #include "vast/query.hpp"
+#include "vast/system/active_partition.hpp"
 #include "vast/system/index.hpp"
 #include "vast/system/meta_index.hpp"
-#include "vast/system/partition.hpp"
+#include "vast/system/passive_partition.hpp"
 #include "vast/system/posix_filesystem.hpp"
 #include "vast/table_slice.hpp"
 #include "vast/table_slice_builder_factory.hpp"
